refactor: refactored theme settings component
