+++ src/components/Toolbar/components/ArrowSettings/index.tsx
@@ -140,7 +140,7 @@ const ArrowSettings: FC<ArrowSettingsProps> = ({
             min={MINIMUM_ARROWHEAD_SCALE}
             max={MAXIMUM_ARROWHEAD_SCALE}
             step="0.1"
-            className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-gray-700"
+            className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-zinc-700 ease duration-200"
           />
         </div>
       </div>
@@ -154,7 +154,7 @@ const ArrowSettings: FC<ArrowSettingsProps> = ({
             min={MINIMUM_ARROWHEAD_SCALE}
             max={MAXIMUM_ARROWHEAD_SCALE}
             step="0.1"
-            className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-gray-700"
+            className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-zinc-700 ease duration-200"
           />
         </div>
       </div>
