+++ src/components/Toolbar/components/MarkerSettings/index.tsx
@@ -201,7 +201,7 @@ const MarkerSettings: FC<MarkerSettingsProps> = ({
             min={MINIMUM_SCALE}
             max={MAXIMUM_SCALE}
             step="0.1"
-            className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-gray-700"
+            className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-zinc-700 ease duration-200"
           />
         </div>
       </div>
@@ -216,7 +216,7 @@ const MarkerSettings: FC<MarkerSettingsProps> = ({
               min={0}
               max={360}
               step="1"
-              className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-gray-700"
+              className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-zinc-700 ease duration-200"
             />
           </div>
         </div>
